run all ut with coverage
