+++ Makefile
@@ -375,7 +375,7 @@ scan:
 
 .PHONY: ut
 ut:
-	ginkgo -mod=mod -progress --always-emit-ginkgo-writer --slow-spec-threshold=60s test/unittest
+	go test -covermode=atomic -count=1 `go list ./... | grep -v test | grep pkg | grep -v daemon`
 
 .PHONY: e2e
 e2e:
