+++ pkg/ipam/ipam_test.go
@@ -2,12 +2,16 @@ package ipam
 
 import (
 	"fmt"
+	"testing"
 
 	. "github.com/onsi/ginkgo/v2"
 	. "github.com/onsi/gomega"
-
-	"github.com/kubeovn/kube-ovn/pkg/ipam"
 )
+
+func TestUTIPAM(t *testing.T) {
+	RegisterFailHandler(Fail)
+	RunSpecs(t, "Kube-OVN ipam unit test Suite")
+}
 
 var _ = Describe("[IPAM]", func() {
 	subnetName := "test"
@@ -34,22 +38,32 @@ var _ = Describe("[IPAM]", func() {
 	dualExcludeIPs := append(ipv4ExcludeIPs, ipv6ExcludeIPs...)
 
 	Describe("[IPAM]", func() {
+		It("subnet not found", func() {
+			im := NewIPAM()
+			Expect(im.GetStaticAddress("pod", "eth0", "", "", subnetName, true)).Error().Should(MatchError(ErrNoAvailable))
+			Expect(im.GetSubnetV4Mask(subnetName)).Error().Should(MatchError(ErrNoAvailable))
+		})
+
 		Context("[IPv4]", func() {
 			It("invalid subnet", func() {
-				im := ipam.NewIPAM()
+				im := NewIPAM()
 				err := im.AddOrUpdateSubnet(subnetName, "1.1.1.1/64", v4Gw, nil)
-				Expect(err).Should(MatchError(ipam.ErrInvalidCIDR))
+				Expect(err).Should(MatchError(ErrInvalidCIDR))
 				err = im.AddOrUpdateSubnet(subnetName, "1.1.256.1/24", v4Gw, nil)
-				Expect(err).Should(MatchError(ipam.ErrInvalidCIDR))
+				Expect(err).Should(MatchError(ErrInvalidCIDR))
+				Expect(im.Subnets[subnetName]).Should(BeNil())
 			})
 
 			It("normal subnet", func() {
-				im := ipam.NewIPAM()
+				im := NewIPAM()
 				err := im.AddOrUpdateSubnet(subnetName, ipv4CIDR, v4Gw, ipv4ExcludeIPs)
 				Expect(err).ShouldNot(HaveOccurred())
+				Expect(im.Subnets[subnetName]).ShouldNot(BeNil())
+				Expect(im.GetSubnetV4Mask(subnetName)).Should(Equal("16"))
 
 				_, _, _, err = im.GetStaticAddress("pod1.ns", "pod1.ns", "10.16.0.2", "", subnetName, true)
 				Expect(err).ShouldNot(HaveOccurred())
+				Expect(im.IsIPAssignedToPod("10.16.0.2", subnetName, "pod1.ns")).Should(BeTrue())
 
 				ip, _, _, err := im.GetRandomAddress("pod1.ns", "pod1.ns", "", subnetName, nil, true)
 				Expect(err).ShouldNot(HaveOccurred())
@@ -60,14 +74,15 @@ var _ = Describe("[IPAM]", func() {
 				Expect(ip).To(Equal("10.16.0.3"))
 
 				_, _, _, err = im.GetStaticAddress("pod3.ns", "pod3.ns", "10.16.0.2", "", subnetName, true)
-				Expect(err).Should(MatchError(ipam.ErrConflict))
+				Expect(err).Should(MatchError(ErrConflict))
 
 				im.ReleaseAddressByPod("pod1.ns")
 				_, _, _, err = im.GetStaticAddress("pod3.ns", "pod3.ns", "10.16.0.2", "", subnetName, true)
 				Expect(err).ShouldNot(HaveOccurred())
+				Expect(im.IsIPAssignedToPod("10.16.0.2", subnetName, "pod1.ns")).Should(BeFalse())
 
 				_, _, _, err = im.GetRandomAddress("pod4.ns", "pod4.ns", "", "invalid_subnet", nil, true)
-				Expect(err).Should(MatchError(ipam.ErrNoAvailable))
+				Expect(err).Should(MatchError(ErrNoAvailable))
 
 				err = im.AddOrUpdateSubnet(subnetName, ipv4CIDR, v4Gw, nil)
 				Expect(err).ShouldNot(HaveOccurred())
@@ -79,10 +94,13 @@ var _ = Describe("[IPAM]", func() {
 				addresses := im.GetPodAddress("pod5.ns")
 				Expect(addresses).To(HaveLen(1))
 				Expect(addresses[0].Ip).To(Equal("10.16.0.1"))
+
+				im.DeleteSubnet(subnetName)
+				Expect(im.Subnets[subnetName]).Should(BeNil())
 			})
 
 			It("change cidr", func() {
-				im := ipam.NewIPAM()
+				im := NewIPAM()
 				err := im.AddOrUpdateSubnet(subnetName, ipv4CIDR, v4Gw, ipv4ExcludeIPs)
 				Expect(err).ShouldNot(HaveOccurred())
 
@@ -91,12 +109,16 @@ var _ = Describe("[IPAM]", func() {
 				ip, _, _, err := im.GetRandomAddress("pod5.ns", "pod5.ns", "", subnetName, nil, true)
 				Expect(err).ShouldNot(HaveOccurred())
 				Expect(ip).To(Equal("10.17.0.2"))
+
+				im.DeleteSubnet(subnetName)
+				Expect(im.Subnets[subnetName]).Should(BeNil())
 			})
 
 			It("reuse released address when no unused address", func() {
-				im := ipam.NewIPAM()
+				im := NewIPAM()
 				err := im.AddOrUpdateSubnet(subnetName, "10.16.0.0/30", v4Gw, nil)
 				Expect(err).ShouldNot(HaveOccurred())
+				Expect(im.Subnets[subnetName]).ShouldNot(BeNil())
 
 				ip, _, _, err := im.GetRandomAddress("pod1.ns", "pod1.ns", "", subnetName, nil, true)
 				Expect(err).ShouldNot(HaveOccurred())
@@ -111,12 +133,16 @@ var _ = Describe("[IPAM]", func() {
 				ip, _, _, err = im.GetRandomAddress("pod1.ns", "pod1.ns", "", subnetName, nil, true)
 				Expect(err).ShouldNot(HaveOccurred())
 				Expect(ip).To(Equal("10.16.0.1"))
+
+				im.DeleteSubnet(subnetName)
+				Expect(im.Subnets[subnetName]).Should(BeNil())
 			})
 
 			It("do not reuse released address after update subnet's excludedIps", func() {
-				im := ipam.NewIPAM()
+				im := NewIPAM()
 				err := im.AddOrUpdateSubnet(subnetName, "10.16.0.0/30", v4Gw, nil)
 				Expect(err).ShouldNot(HaveOccurred())
+				Expect(im.Subnets[subnetName]).ShouldNot(BeNil())
 
 				ip, _, _, err := im.GetRandomAddress("pod1.ns", "pod1.ns", "", subnetName, nil, true)
 				Expect(err).ShouldNot(HaveOccurred())
@@ -127,23 +153,28 @@ var _ = Describe("[IPAM]", func() {
 				Expect(err).ShouldNot(HaveOccurred())
 
 				_, _, _, err = im.GetRandomAddress("pod1.ns", "pod1.ns", "", subnetName, nil, true)
-				Expect(err).Should(MatchError(ipam.ErrNoAvailable))
+				Expect(err).Should(MatchError(ErrNoAvailable))
+
+				im.DeleteSubnet(subnetName)
+				Expect(im.Subnets[subnetName]).Should(BeNil())
 			})
 		})
 
 		Context("[IPv6]", func() {
 			It("invalid subnet", func() {
-				im := ipam.NewIPAM()
+				im := NewIPAM()
 				err := im.AddOrUpdateSubnet(subnetName, "fd00::/130", v6Gw, nil)
-				Expect(err).Should(MatchError(ipam.ErrInvalidCIDR))
+				Expect(err).Should(MatchError(ErrInvalidCIDR))
 				err = im.AddOrUpdateSubnet(subnetName, "fd00::g/120", v6Gw, nil)
-				Expect(err).Should(MatchError(ipam.ErrInvalidCIDR))
+				Expect(err).Should(MatchError(ErrInvalidCIDR))
+				Expect(im.Subnets[subnetName]).Should(BeNil())
 			})
 
 			It("normal subnet", func() {
-				im := ipam.NewIPAM()
+				im := NewIPAM()
 				err := im.AddOrUpdateSubnet(subnetName, ipv6CIDR, v6Gw, ipv6ExcludeIPs)
 				Expect(err).ShouldNot(HaveOccurred())
+				Expect(im.Subnets[subnetName]).ShouldNot(BeNil())
 
 				_, _, _, err = im.GetStaticAddress("pod1.ns", "pod1.ns", "fd00::2", "", subnetName, true)
 				Expect(err).ShouldNot(HaveOccurred())
@@ -157,14 +188,14 @@ var _ = Describe("[IPAM]", func() {
 				Expect(ip).To(Equal("fd00::3"))
 
 				_, _, _, err = im.GetStaticAddress("pod3.ns", "pod3.ns", "fd00::2", "", subnetName, true)
-				Expect(err).Should(MatchError(ipam.ErrConflict))
+				Expect(err).Should(MatchError(ErrConflict))
 
 				im.ReleaseAddressByPod("pod1.ns")
 				_, _, _, err = im.GetStaticAddress("pod3.ns", "pod3.ns", "fd00::2", "", subnetName, true)
 				Expect(err).ShouldNot(HaveOccurred())
 
 				_, _, _, err = im.GetRandomAddress("pod4.ns", "pod4.ns", "", "invalid_subnet", nil, true)
-				Expect(err).Should(MatchError(ipam.ErrNoAvailable))
+				Expect(err).Should(MatchError(ErrNoAvailable))
 
 				err = im.AddOrUpdateSubnet(subnetName, ipv6CIDR, v6Gw, nil)
 				Expect(err).ShouldNot(HaveOccurred())
@@ -179,9 +210,10 @@ var _ = Describe("[IPAM]", func() {
 			})
 
 			It("change cidr", func() {
-				im := ipam.NewIPAM()
+				im := NewIPAM()
 				err := im.AddOrUpdateSubnet(subnetName, ipv6CIDR, v6Gw, ipv6ExcludeIPs)
 				Expect(err).ShouldNot(HaveOccurred())
+				Expect(im.Subnets[subnetName]).ShouldNot(BeNil())
 
 				err = im.AddOrUpdateSubnet(subnetName, "fe00::/112", v6Gw, []string{"fe00::1"})
 				Expect(err).ShouldNot(HaveOccurred())
@@ -191,9 +223,10 @@ var _ = Describe("[IPAM]", func() {
 			})
 
 			It("reuse released address when no unused address", func() {
-				im := ipam.NewIPAM()
+				im := NewIPAM()
 				err := im.AddOrUpdateSubnet(subnetName, "fd00::/126", v6Gw, nil)
 				Expect(err).ShouldNot(HaveOccurred())
+				Expect(im.Subnets[subnetName]).ShouldNot(BeNil())
 
 				_, ip, _, err := im.GetRandomAddress("pod1.ns", "pod1.ns", "", subnetName, nil, true)
 				Expect(err).ShouldNot(HaveOccurred())
@@ -211,9 +244,10 @@ var _ = Describe("[IPAM]", func() {
 			})
 
 			It("do not reuse released address after update subnet's excludedIps", func() {
-				im := ipam.NewIPAM()
+				im := NewIPAM()
 				err := im.AddOrUpdateSubnet(subnetName, "fd00::/126", v6Gw, nil)
 				Expect(err).ShouldNot(HaveOccurred())
+				Expect(im.Subnets[subnetName]).ShouldNot(BeNil())
 
 				_, ip, _, err := im.GetRandomAddress("pod1.ns", "pod1.ns", "", subnetName, nil, true)
 				Expect(err).ShouldNot(HaveOccurred())
@@ -224,27 +258,30 @@ var _ = Describe("[IPAM]", func() {
 				Expect(err).ShouldNot(HaveOccurred())
 
 				_, _, _, err = im.GetRandomAddress("pod1.ns", "pod1.ns", "", subnetName, nil, true)
-				Expect(err).Should(MatchError(ipam.ErrNoAvailable))
+				Expect(err).Should(MatchError(ErrNoAvailable))
 			})
 		})
 
 		Context("[DualStack]", func() {
 			It("invalid subnet", func() {
-				im := ipam.NewIPAM()
+				im := NewIPAM()
 				err := im.AddOrUpdateSubnet(subnetName, fmt.Sprintf("1.1.1.1/64,%s", ipv6CIDR), dualGw, nil)
-				Expect(err).Should(MatchError(ipam.ErrInvalidCIDR))
+				Expect(err).Should(MatchError(ErrInvalidCIDR))
 				err = im.AddOrUpdateSubnet(subnetName, fmt.Sprintf("1.1.256.1/24,%s", ipv6CIDR), dualGw, nil)
-				Expect(err).Should(MatchError(ipam.ErrInvalidCIDR))
+				Expect(err).Should(MatchError(ErrInvalidCIDR))
 				err = im.AddOrUpdateSubnet(subnetName, fmt.Sprintf("%s,fd00::/130", ipv4CIDR), dualGw, nil)
-				Expect(err).Should(MatchError(ipam.ErrInvalidCIDR))
+				Expect(err).Should(MatchError(ErrInvalidCIDR))
 				err = im.AddOrUpdateSubnet(subnetName, fmt.Sprintf("%s,fd00::g/120", ipv4CIDR), dualGw, nil)
-				Expect(err).Should(MatchError(ipam.ErrInvalidCIDR))
+				Expect(err).Should(MatchError(ErrInvalidCIDR))
+				Expect(im.Subnets[subnetName]).Should(BeNil())
 			})
 
 			It("normal subnet", func() {
-				im := ipam.NewIPAM()
+				im := NewIPAM()
 				err := im.AddOrUpdateSubnet(subnetName, dualCIDR, dualGw, dualExcludeIPs)
 				Expect(err).ShouldNot(HaveOccurred())
+				Expect(im.Subnets[subnetName]).ShouldNot(BeNil())
+				Expect(im.GetSubnetV4Mask(subnetName)).Should(Equal("16"))
 
 				_, _, _, err = im.GetStaticAddress("pod1.ns", "pod1.ns", "10.16.0.2,fd00::2", "", subnetName, true)
 				Expect(err).ShouldNot(HaveOccurred())
@@ -260,14 +297,14 @@ var _ = Describe("[IPAM]", func() {
 				Expect(ipv6).To(Equal("fd00::3"))
 
 				_, _, _, err = im.GetStaticAddress("pod3.ns", "pod3.ns", "10.16.0.2,fd00::2", "", subnetName, true)
-				Expect(err).Should(MatchError(ipam.ErrConflict))
+				Expect(err).Should(MatchError(ErrConflict))
 
 				im.ReleaseAddressByPod("pod1.ns")
 				_, _, _, err = im.GetStaticAddress("pod3.ns", "pod3.ns", "10.16.0.2,fd00::2", "", subnetName, true)
 				Expect(err).ShouldNot(HaveOccurred())
 
 				_, _, _, err = im.GetRandomAddress("pod4.ns", "pod4.ns", "", "invalid_subnet", nil, true)
-				Expect(err).Should(MatchError(ipam.ErrNoAvailable))
+				Expect(err).Should(MatchError(ErrNoAvailable))
 
 				err = im.AddOrUpdateSubnet(subnetName, dualCIDR, dualGw, nil)
 				Expect(err).ShouldNot(HaveOccurred())
@@ -284,9 +321,10 @@ var _ = Describe("[IPAM]", func() {
 			})
 
 			It("change cidr", func() {
-				im := ipam.NewIPAM()
+				im := NewIPAM()
 				err := im.AddOrUpdateSubnet(subnetName, dualCIDR, dualGw, dualExcludeIPs)
 				Expect(err).ShouldNot(HaveOccurred())
+				Expect(im.Subnets[subnetName]).ShouldNot(BeNil())
 
 				err = im.AddOrUpdateSubnet(subnetName, "10.17.0.2/16,fe00::/112", dualGw, []string{"10.17.0.1", "fe00::1"})
 				Expect(err).ShouldNot(HaveOccurred())
@@ -297,9 +335,10 @@ var _ = Describe("[IPAM]", func() {
 			})
 
 			It("reuse released address when no unused address", func() {
-				im := ipam.NewIPAM()
+				im := NewIPAM()
 				err := im.AddOrUpdateSubnet(subnetName, "10.16.0.2/30,fd00::/126", dualGw, nil)
 				Expect(err).ShouldNot(HaveOccurred())
+				Expect(im.Subnets[subnetName]).ShouldNot(BeNil())
 
 				ipv4, ipv6, _, err := im.GetRandomAddress("pod1.ns", "pod1.ns", "", subnetName, nil, true)
 				Expect(err).ShouldNot(HaveOccurred())
@@ -320,9 +359,10 @@ var _ = Describe("[IPAM]", func() {
 			})
 
 			It("do not reuse released address after update subnet's excludedIps", func() {
-				im := ipam.NewIPAM()
+				im := NewIPAM()
 				err := im.AddOrUpdateSubnet(subnetName, "10.16.0.2/30,fd00::/126", dualGw, nil)
 				Expect(err).ShouldNot(HaveOccurred())
+				Expect(im.Subnets[subnetName]).ShouldNot(BeNil())
 
 				ipv4, ipv6, _, err := im.GetRandomAddress("pod1.ns", "pod1.ns", "", subnetName, nil, true)
 				Expect(err).ShouldNot(HaveOccurred())
@@ -334,15 +374,15 @@ var _ = Describe("[IPAM]", func() {
 				Expect(err).ShouldNot(HaveOccurred())
 
 				_, _, _, err = im.GetRandomAddress("pod1.ns", "pod1.ns", "", subnetName, nil, true)
-				Expect(err).Should(MatchError(ipam.ErrNoAvailable))
+				Expect(err).Should(MatchError(ErrNoAvailable))
 			})
 		})
 	})
 
 	Describe("[IP]", func() {
 		It("IPv4 operation", func() {
-			ip1 := ipam.IP("10.0.0.16")
-			ip2 := ipam.IP("10.0.0.17")
+			ip1 := IP("10.0.0.16")
+			ip2 := IP("10.0.0.17")
 
 			Expect(ip1.Equal(ip1)).To(BeTrue())
 			Expect(ip1.GreaterThan(ip1)).To(BeFalse())
@@ -359,8 +399,8 @@ var _ = Describe("[IPAM]", func() {
 		})
 
 		It("IPv6 operation", func() {
-			ip1 := ipam.IP("fd00::16")
-			ip2 := ipam.IP("fd00::17")
+			ip1 := IP("fd00::16")
+			ip2 := IP("fd00::17")
 
 			Expect(ip1.Equal(ip1)).To(BeTrue())
 			Expect(ip1.GreaterThan(ip1)).To(BeFalse())
@@ -380,21 +420,21 @@ var _ = Describe("[IPAM]", func() {
 	Describe("[Subnet]", func() {
 		Context("[IPv4]", func() {
 			It("init subnet", func() {
-				subnet, err := ipam.NewSubnet(subnetName, ipv4CIDR, ipv4ExcludeIPs)
+				subnet, err := NewSubnet(subnetName, ipv4CIDR, ipv4ExcludeIPs)
 				Expect(err).ShouldNot(HaveOccurred())
 				Expect(subnet.Name).To(Equal(subnetName))
 				Expect(subnet.V4ReservedIPList).To(HaveLen(len(ipv4ExcludeIPs) - 1))
 				Expect(subnet.V4FreeIPList).To(HaveLen(3))
 				Expect(subnet.V4FreeIPList).To(Equal(
-					ipam.IPRangeList{
-						&ipam.IPRange{Start: "10.16.0.2", End: "10.16.0.3"},
-						&ipam.IPRange{Start: "10.16.0.5", End: "10.16.0.9"},
-						&ipam.IPRange{Start: "10.16.0.24", End: "10.16.255.254"},
+					IPRangeList{
+						&IPRange{Start: "10.16.0.2", End: "10.16.0.3"},
+						&IPRange{Start: "10.16.0.5", End: "10.16.0.9"},
+						&IPRange{Start: "10.16.0.24", End: "10.16.255.254"},
 					}))
 			})
 
 			It("static allocation", func() {
-				subnet, err := ipam.NewSubnet(subnetName, ipv4CIDR, ipv4ExcludeIPs)
+				subnet, err := NewSubnet(subnetName, ipv4CIDR, ipv4ExcludeIPs)
 				Expect(err).ShouldNot(HaveOccurred())
 				_, _, err = subnet.GetStaticAddress("pod1.ns", "pod1.ns", "10.16.0.2", "", false, true)
 				Expect(err).ShouldNot(HaveOccurred())
@@ -403,30 +443,30 @@ var _ = Describe("[IPAM]", func() {
 				_, _, err = subnet.GetStaticAddress("pod3.ns", "pod3.ns", "10.16.0.20", "", false, true)
 				Expect(err).ShouldNot(HaveOccurred())
 				Expect(subnet.V4FreeIPList).To(Equal(
-					ipam.IPRangeList{
-						&ipam.IPRange{Start: "10.16.0.5", End: "10.16.0.9"},
-						&ipam.IPRange{Start: "10.16.0.24", End: "10.16.255.254"},
+					IPRangeList{
+						&IPRange{Start: "10.16.0.5", End: "10.16.0.9"},
+						&IPRange{Start: "10.16.0.24", End: "10.16.255.254"},
 					}))
 
-				Expect(subnet.V4IPToPod).To(HaveKeyWithValue(ipam.IP("10.16.0.2"), "pod1.ns"))
-				Expect(subnet.V4IPToPod).To(HaveKeyWithValue(ipam.IP("10.16.0.3"), "pod2.ns"))
-				Expect(subnet.V4IPToPod).To(HaveKeyWithValue(ipam.IP("10.16.0.20"), "pod3.ns"))
-				Expect(subnet.V4NicToIP).To(HaveKeyWithValue("pod1.ns", ipam.IP("10.16.0.2")))
-				Expect(subnet.V4NicToIP).To(HaveKeyWithValue("pod2.ns", ipam.IP("10.16.0.3")))
-				Expect(subnet.V4NicToIP).To(HaveKeyWithValue("pod3.ns", ipam.IP("10.16.0.20")))
+				Expect(subnet.V4IPToPod).To(HaveKeyWithValue(IP("10.16.0.2"), "pod1.ns"))
+				Expect(subnet.V4IPToPod).To(HaveKeyWithValue(IP("10.16.0.3"), "pod2.ns"))
+				Expect(subnet.V4IPToPod).To(HaveKeyWithValue(IP("10.16.0.20"), "pod3.ns"))
+				Expect(subnet.V4NicToIP).To(HaveKeyWithValue("pod1.ns", IP("10.16.0.2")))
+				Expect(subnet.V4NicToIP).To(HaveKeyWithValue("pod2.ns", IP("10.16.0.3")))
+				Expect(subnet.V4NicToIP).To(HaveKeyWithValue("pod3.ns", IP("10.16.0.20")))
 
 				_, _, err = subnet.GetStaticAddress("pod4.ns", "pod4.ns", "10.16.0.3", "", false, true)
-				Expect(err).Should(MatchError(ipam.ErrConflict))
+				Expect(err).Should(MatchError(ErrConflict))
 				_, _, err = subnet.GetStaticAddress("pod5.ns", "pod5.ns", "19.16.0.3", "", false, true)
-				Expect(err).Should(MatchError(ipam.ErrOutOfRange))
+				Expect(err).Should(MatchError(ErrOutOfRange))
 
 				subnet.ReleaseAddress("pod1.ns")
 				subnet.ReleaseAddress("pod2.ns")
 				subnet.ReleaseAddress("pod3.ns")
 				Expect(subnet.V4FreeIPList).To(Equal(
-					ipam.IPRangeList{
-						&ipam.IPRange{Start: "10.16.0.5", End: "10.16.0.9"},
-						&ipam.IPRange{Start: "10.16.0.24", End: "10.16.255.254"},
+					IPRangeList{
+						&IPRange{Start: "10.16.0.5", End: "10.16.0.9"},
+						&IPRange{Start: "10.16.0.24", End: "10.16.255.254"},
 					}))
 
 				Expect(subnet.V4NicToIP).To(BeEmpty())
@@ -434,33 +474,33 @@ var _ = Describe("[IPAM]", func() {
 			})
 
 			It("random allocation", func() {
-				subnet, err := ipam.NewSubnet(subnetName, "10.16.0.0/30", nil)
+				subnet, err := NewSubnet(subnetName, "10.16.0.0/30", nil)
 				Expect(err).ShouldNot(HaveOccurred())
 
 				ip1, _, _, err := subnet.GetRandomAddress("pod1.ns", "pod1.ns", "", nil, true)
 				Expect(err).ShouldNot(HaveOccurred())
-				Expect(ip1).To(Equal(ipam.IP("10.16.0.1")))
+				Expect(ip1).To(Equal(IP("10.16.0.1")))
 				ip1, _, _, err = subnet.GetRandomAddress("pod1.ns", "pod1.ns", "", nil, true)
 				Expect(err).ShouldNot(HaveOccurred())
-				Expect(ip1).To(Equal(ipam.IP("10.16.0.1")))
+				Expect(ip1).To(Equal(IP("10.16.0.1")))
 
 				ip2, _, _, err := subnet.GetRandomAddress("pod2.ns", "pod2.ns", "", nil, true)
 				Expect(err).ShouldNot(HaveOccurred())
-				Expect(ip2).To(Equal(ipam.IP("10.16.0.2")))
+				Expect(ip2).To(Equal(IP("10.16.0.2")))
 
 				_, _, _, err = subnet.GetRandomAddress("pod3.ns", "pod3.ns", "", nil, true)
-				Expect(err).Should(MatchError(ipam.ErrNoAvailable))
+				Expect(err).Should(MatchError(ErrNoAvailable))
 				Expect(subnet.V4FreeIPList).To(BeEmpty())
 
-				Expect(subnet.V4IPToPod).To(HaveKeyWithValue(ipam.IP("10.16.0.1"), "pod1.ns"))
-				Expect(subnet.V4IPToPod).To(HaveKeyWithValue(ipam.IP("10.16.0.2"), "pod2.ns"))
-				Expect(subnet.V4NicToIP).To(HaveKeyWithValue("pod1.ns", ipam.IP("10.16.0.1")))
-				Expect(subnet.V4NicToIP).To(HaveKeyWithValue("pod2.ns", ipam.IP("10.16.0.2")))
+				Expect(subnet.V4IPToPod).To(HaveKeyWithValue(IP("10.16.0.1"), "pod1.ns"))
+				Expect(subnet.V4IPToPod).To(HaveKeyWithValue(IP("10.16.0.2"), "pod2.ns"))
+				Expect(subnet.V4NicToIP).To(HaveKeyWithValue("pod1.ns", IP("10.16.0.1")))
+				Expect(subnet.V4NicToIP).To(HaveKeyWithValue("pod2.ns", IP("10.16.0.2")))
 
 				subnet.ReleaseAddress("pod1.ns")
 				subnet.ReleaseAddress("pod2.ns")
-				Expect(subnet.V4FreeIPList).To(Equal(ipam.IPRangeList{}))
-				Expect(subnet.V4ReleasedIPList).To(Equal(ipam.IPRangeList{&ipam.IPRange{Start: "10.16.0.1", End: "10.16.0.2"}}))
+				Expect(subnet.V4FreeIPList).To(Equal(IPRangeList{}))
+				Expect(subnet.V4ReleasedIPList).To(Equal(IPRangeList{&IPRange{Start: "10.16.0.1", End: "10.16.0.2"}}))
 				Expect(subnet.V4IPToPod).To(BeEmpty())
 				Expect(subnet.V4NicToIP).To(BeEmpty())
 			})
@@ -468,21 +508,21 @@ var _ = Describe("[IPAM]", func() {
 
 		Context("[IPv6]", func() {
 			It("init subnet", func() {
-				subnet, err := ipam.NewSubnet(subnetName, ipv6CIDR, ipv6ExcludeIPs)
+				subnet, err := NewSubnet(subnetName, ipv6CIDR, ipv6ExcludeIPs)
 				Expect(err).ShouldNot(HaveOccurred())
 				Expect(subnet.Name).To(Equal(subnetName))
 				Expect(subnet.V6ReservedIPList).To(HaveLen(len(ipv6ExcludeIPs) - 1))
 				Expect(subnet.V6FreeIPList).To(HaveLen(3))
 				Expect(subnet.V6FreeIPList).To(Equal(
-					ipam.IPRangeList{
-						&ipam.IPRange{Start: "fd00::2", End: "fd00::3"},
-						&ipam.IPRange{Start: "fd00::5", End: "fd00::9"},
-						&ipam.IPRange{Start: "fd00::18", End: "fd00::fffe"},
+					IPRangeList{
+						&IPRange{Start: "fd00::2", End: "fd00::3"},
+						&IPRange{Start: "fd00::5", End: "fd00::9"},
+						&IPRange{Start: "fd00::18", End: "fd00::fffe"},
 					}))
 			})
 
 			It("static allocation", func() {
-				subnet, err := ipam.NewSubnet(subnetName, ipv6CIDR, ipv6ExcludeIPs)
+				subnet, err := NewSubnet(subnetName, ipv6CIDR, ipv6ExcludeIPs)
 				Expect(err).ShouldNot(HaveOccurred())
 				_, _, err = subnet.GetStaticAddress("pod1.ns", "pod1.ns", "fd00::2", "", false, true)
 				Expect(err).ShouldNot(HaveOccurred())
@@ -491,30 +531,30 @@ var _ = Describe("[IPAM]", func() {
 				_, _, err = subnet.GetStaticAddress("pod3.ns", "pod3.ns", "fd00::14", "", false, true)
 				Expect(err).ShouldNot(HaveOccurred())
 				Expect(subnet.V6FreeIPList).To(Equal(
-					ipam.IPRangeList{
-						&ipam.IPRange{Start: "fd00::5", End: "fd00::9"},
-						&ipam.IPRange{Start: "fd00::18", End: "fd00::fffe"},
+					IPRangeList{
+						&IPRange{Start: "fd00::5", End: "fd00::9"},
+						&IPRange{Start: "fd00::18", End: "fd00::fffe"},
 					}))
 
-				Expect(subnet.V6IPToPod).To(HaveKeyWithValue(ipam.IP("fd00::2"), "pod1.ns"))
-				Expect(subnet.V6IPToPod).To(HaveKeyWithValue(ipam.IP("fd00::3"), "pod2.ns"))
-				Expect(subnet.V6IPToPod).To(HaveKeyWithValue(ipam.IP("fd00::14"), "pod3.ns"))
-				Expect(subnet.V6NicToIP).To(HaveKeyWithValue("pod1.ns", ipam.IP("fd00::2")))
-				Expect(subnet.V6NicToIP).To(HaveKeyWithValue("pod2.ns", ipam.IP("fd00::3")))
-				Expect(subnet.V6NicToIP).To(HaveKeyWithValue("pod3.ns", ipam.IP("fd00::14")))
+				Expect(subnet.V6IPToPod).To(HaveKeyWithValue(IP("fd00::2"), "pod1.ns"))
+				Expect(subnet.V6IPToPod).To(HaveKeyWithValue(IP("fd00::3"), "pod2.ns"))
+				Expect(subnet.V6IPToPod).To(HaveKeyWithValue(IP("fd00::14"), "pod3.ns"))
+				Expect(subnet.V6NicToIP).To(HaveKeyWithValue("pod1.ns", IP("fd00::2")))
+				Expect(subnet.V6NicToIP).To(HaveKeyWithValue("pod2.ns", IP("fd00::3")))
+				Expect(subnet.V6NicToIP).To(HaveKeyWithValue("pod3.ns", IP("fd00::14")))
 
 				_, _, err = subnet.GetStaticAddress("pod4.ns", "pod4.ns", "fd00::3", "", false, true)
-				Expect(err).Should(MatchError(ipam.ErrConflict))
+				Expect(err).Should(MatchError(ErrConflict))
 				_, _, err = subnet.GetStaticAddress("pod5.ns", "pod5.ns", "fe00::3", "", false, true)
-				Expect(err).Should(MatchError(ipam.ErrOutOfRange))
+				Expect(err).Should(MatchError(ErrOutOfRange))
 
 				subnet.ReleaseAddress("pod1.ns")
 				subnet.ReleaseAddress("pod2.ns")
 				subnet.ReleaseAddress("pod3.ns")
 				Expect(subnet.V6FreeIPList).To(Equal(
-					ipam.IPRangeList{
-						&ipam.IPRange{Start: "fd00::5", End: "fd00::9"},
-						&ipam.IPRange{Start: "fd00::18", End: "fd00::fffe"},
+					IPRangeList{
+						&IPRange{Start: "fd00::5", End: "fd00::9"},
+						&IPRange{Start: "fd00::18", End: "fd00::fffe"},
 					}))
 
 				Expect(subnet.V6NicToIP).To(BeEmpty())
@@ -522,33 +562,33 @@ var _ = Describe("[IPAM]", func() {
 			})
 
 			It("random allocation", func() {
-				subnet, err := ipam.NewSubnet(subnetName, "fd00::/126", nil)
+				subnet, err := NewSubnet(subnetName, "fd00::/126", nil)
 				Expect(err).ShouldNot(HaveOccurred())
 
 				_, ip1, _, err := subnet.GetRandomAddress("pod1.ns", "pod1.ns", "", nil, true)
 				Expect(err).ShouldNot(HaveOccurred())
-				Expect(ip1).To(Equal(ipam.IP("fd00::1")))
+				Expect(ip1).To(Equal(IP("fd00::1")))
 				_, ip1, _, err = subnet.GetRandomAddress("pod1.ns", "pod1.ns", "", nil, true)
 				Expect(err).ShouldNot(HaveOccurred())
-				Expect(ip1).To(Equal(ipam.IP("fd00::1")))
+				Expect(ip1).To(Equal(IP("fd00::1")))
 
 				_, ip2, _, err := subnet.GetRandomAddress("pod2.ns", "pod2.ns", "", nil, true)
 				Expect(err).ShouldNot(HaveOccurred())
-				Expect(ip2).To(Equal(ipam.IP("fd00::2")))
+				Expect(ip2).To(Equal(IP("fd00::2")))
 
 				_, _, _, err = subnet.GetRandomAddress("pod3.ns", "pod3.ns", "", nil, true)
-				Expect(err).Should(MatchError(ipam.ErrNoAvailable))
+				Expect(err).Should(MatchError(ErrNoAvailable))
 				Expect(subnet.V6FreeIPList).To(BeEmpty())
 
-				Expect(subnet.V6IPToPod).To(HaveKeyWithValue(ipam.IP("fd00::1"), "pod1.ns"))
-				Expect(subnet.V6IPToPod).To(HaveKeyWithValue(ipam.IP("fd00::2"), "pod2.ns"))
-				Expect(subnet.V6NicToIP).To(HaveKeyWithValue("pod1.ns", ipam.IP("fd00::1")))
-				Expect(subnet.V6NicToIP).To(HaveKeyWithValue("pod2.ns", ipam.IP("fd00::2")))
+				Expect(subnet.V6IPToPod).To(HaveKeyWithValue(IP("fd00::1"), "pod1.ns"))
+				Expect(subnet.V6IPToPod).To(HaveKeyWithValue(IP("fd00::2"), "pod2.ns"))
+				Expect(subnet.V6NicToIP).To(HaveKeyWithValue("pod1.ns", IP("fd00::1")))
+				Expect(subnet.V6NicToIP).To(HaveKeyWithValue("pod2.ns", IP("fd00::2")))
 
 				subnet.ReleaseAddress("pod1.ns")
 				subnet.ReleaseAddress("pod2.ns")
-				Expect(subnet.V6FreeIPList).To(Equal(ipam.IPRangeList{}))
-				Expect(subnet.V6ReleasedIPList).To(Equal(ipam.IPRangeList{&ipam.IPRange{Start: "fd00::1", End: "fd00::2"}}))
+				Expect(subnet.V6FreeIPList).To(Equal(IPRangeList{}))
+				Expect(subnet.V6ReleasedIPList).To(Equal(IPRangeList{&IPRange{Start: "fd00::1", End: "fd00::2"}}))
 				Expect(subnet.V6IPToPod).To(BeEmpty())
 				Expect(subnet.V6NicToIP).To(BeEmpty())
 			})
@@ -556,29 +596,29 @@ var _ = Describe("[IPAM]", func() {
 
 		Context("[DualStack]", func() {
 			It("init subnet", func() {
-				subnet, err := ipam.NewSubnet(subnetName, dualCIDR, dualExcludeIPs)
+				subnet, err := NewSubnet(subnetName, dualCIDR, dualExcludeIPs)
 				Expect(err).ShouldNot(HaveOccurred())
 				Expect(subnet.Name).To(Equal(subnetName))
 				Expect(subnet.V4ReservedIPList).To(HaveLen(len(ipv4ExcludeIPs) - 1))
 				Expect(subnet.V4FreeIPList).To(HaveLen(3))
 				Expect(subnet.V4FreeIPList).To(Equal(
-					ipam.IPRangeList{
-						&ipam.IPRange{Start: "10.16.0.2", End: "10.16.0.3"},
-						&ipam.IPRange{Start: "10.16.0.5", End: "10.16.0.9"},
-						&ipam.IPRange{Start: "10.16.0.24", End: "10.16.255.254"},
+					IPRangeList{
+						&IPRange{Start: "10.16.0.2", End: "10.16.0.3"},
+						&IPRange{Start: "10.16.0.5", End: "10.16.0.9"},
+						&IPRange{Start: "10.16.0.24", End: "10.16.255.254"},
 					}))
 				Expect(subnet.V6ReservedIPList).To(HaveLen(len(ipv6ExcludeIPs) - 1))
 				Expect(subnet.V6FreeIPList).To(HaveLen(3))
 				Expect(subnet.V6FreeIPList).To(Equal(
-					ipam.IPRangeList{
-						&ipam.IPRange{Start: "fd00::2", End: "fd00::3"},
-						&ipam.IPRange{Start: "fd00::5", End: "fd00::9"},
-						&ipam.IPRange{Start: "fd00::18", End: "fd00::fffe"},
+					IPRangeList{
+						&IPRange{Start: "fd00::2", End: "fd00::3"},
+						&IPRange{Start: "fd00::5", End: "fd00::9"},
+						&IPRange{Start: "fd00::18", End: "fd00::fffe"},
 					}))
 			})
 
 			It("static allocation", func() {
-				subnet, err := ipam.NewSubnet(subnetName, dualCIDR, dualExcludeIPs)
+				subnet, err := NewSubnet(subnetName, dualCIDR, dualExcludeIPs)
 				Expect(err).ShouldNot(HaveOccurred())
 				_, _, err = subnet.GetStaticAddress("pod1.ns", "pod1.ns", "10.16.0.2", "", false, true)
 				Expect(err).ShouldNot(HaveOccurred())
@@ -594,50 +634,50 @@ var _ = Describe("[IPAM]", func() {
 				Expect(err).ShouldNot(HaveOccurred())
 
 				Expect(subnet.V4FreeIPList).To(Equal(
-					ipam.IPRangeList{
-						&ipam.IPRange{Start: "10.16.0.5", End: "10.16.0.9"},
-						&ipam.IPRange{Start: "10.16.0.24", End: "10.16.255.254"},
+					IPRangeList{
+						&IPRange{Start: "10.16.0.5", End: "10.16.0.9"},
+						&IPRange{Start: "10.16.0.24", End: "10.16.255.254"},
 					}))
 				Expect(subnet.V6FreeIPList).To(Equal(
-					ipam.IPRangeList{
-						&ipam.IPRange{Start: "fd00::5", End: "fd00::9"},
-						&ipam.IPRange{Start: "fd00::18", End: "fd00::fffe"},
+					IPRangeList{
+						&IPRange{Start: "fd00::5", End: "fd00::9"},
+						&IPRange{Start: "fd00::18", End: "fd00::fffe"},
 					}))
 
-				Expect(subnet.V4IPToPod).To(HaveKeyWithValue(ipam.IP("10.16.0.2"), "pod1.ns"))
-				Expect(subnet.V4IPToPod).To(HaveKeyWithValue(ipam.IP("10.16.0.3"), "pod2.ns"))
-				Expect(subnet.V4IPToPod).To(HaveKeyWithValue(ipam.IP("10.16.0.20"), "pod3.ns"))
-				Expect(subnet.V4NicToIP).To(HaveKeyWithValue("pod1.ns", ipam.IP("10.16.0.2")))
-				Expect(subnet.V4NicToIP).To(HaveKeyWithValue("pod2.ns", ipam.IP("10.16.0.3")))
-				Expect(subnet.V4NicToIP).To(HaveKeyWithValue("pod3.ns", ipam.IP("10.16.0.20")))
-				Expect(subnet.V6IPToPod).To(HaveKeyWithValue(ipam.IP("fd00::2"), "pod1.ns"))
-				Expect(subnet.V6IPToPod).To(HaveKeyWithValue(ipam.IP("fd00::3"), "pod2.ns"))
-				Expect(subnet.V6IPToPod).To(HaveKeyWithValue(ipam.IP("fd00::14"), "pod3.ns"))
-				Expect(subnet.V6NicToIP).To(HaveKeyWithValue("pod1.ns", ipam.IP("fd00::2")))
-				Expect(subnet.V6NicToIP).To(HaveKeyWithValue("pod2.ns", ipam.IP("fd00::3")))
-				Expect(subnet.V6NicToIP).To(HaveKeyWithValue("pod3.ns", ipam.IP("fd00::14")))
+				Expect(subnet.V4IPToPod).To(HaveKeyWithValue(IP("10.16.0.2"), "pod1.ns"))
+				Expect(subnet.V4IPToPod).To(HaveKeyWithValue(IP("10.16.0.3"), "pod2.ns"))
+				Expect(subnet.V4IPToPod).To(HaveKeyWithValue(IP("10.16.0.20"), "pod3.ns"))
+				Expect(subnet.V4NicToIP).To(HaveKeyWithValue("pod1.ns", IP("10.16.0.2")))
+				Expect(subnet.V4NicToIP).To(HaveKeyWithValue("pod2.ns", IP("10.16.0.3")))
+				Expect(subnet.V4NicToIP).To(HaveKeyWithValue("pod3.ns", IP("10.16.0.20")))
+				Expect(subnet.V6IPToPod).To(HaveKeyWithValue(IP("fd00::2"), "pod1.ns"))
+				Expect(subnet.V6IPToPod).To(HaveKeyWithValue(IP("fd00::3"), "pod2.ns"))
+				Expect(subnet.V6IPToPod).To(HaveKeyWithValue(IP("fd00::14"), "pod3.ns"))
+				Expect(subnet.V6NicToIP).To(HaveKeyWithValue("pod1.ns", IP("fd00::2")))
+				Expect(subnet.V6NicToIP).To(HaveKeyWithValue("pod2.ns", IP("fd00::3")))
+				Expect(subnet.V6NicToIP).To(HaveKeyWithValue("pod3.ns", IP("fd00::14")))
 
 				_, _, err = subnet.GetStaticAddress("pod4.ns", "pod4.ns", "10.16.0.3", "", false, true)
-				Expect(err).Should(MatchError(ipam.ErrConflict))
+				Expect(err).Should(MatchError(ErrConflict))
 				_, _, err = subnet.GetStaticAddress("pod4.ns", "pod4.ns", "fd00::3", "", false, true)
-				Expect(err).Should(MatchError(ipam.ErrConflict))
+				Expect(err).Should(MatchError(ErrConflict))
 				_, _, err = subnet.GetStaticAddress("pod5.ns", "pod5.ns", "19.16.0.3", "", false, true)
-				Expect(err).Should(MatchError(ipam.ErrOutOfRange))
+				Expect(err).Should(MatchError(ErrOutOfRange))
 				_, _, err = subnet.GetStaticAddress("pod1.ns", "pod5.ns", "fe00::3", "", false, true)
-				Expect(err).Should(MatchError(ipam.ErrOutOfRange))
+				Expect(err).Should(MatchError(ErrOutOfRange))
 
 				subnet.ReleaseAddress("pod1.ns")
 				subnet.ReleaseAddress("pod2.ns")
 				subnet.ReleaseAddress("pod3.ns")
 				Expect(subnet.V4FreeIPList).To(Equal(
-					ipam.IPRangeList{
-						&ipam.IPRange{Start: "10.16.0.5", End: "10.16.0.9"},
-						&ipam.IPRange{Start: "10.16.0.24", End: "10.16.255.254"},
+					IPRangeList{
+						&IPRange{Start: "10.16.0.5", End: "10.16.0.9"},
+						&IPRange{Start: "10.16.0.24", End: "10.16.255.254"},
 					}))
 				Expect(subnet.V6FreeIPList).To(Equal(
-					ipam.IPRangeList{
-						&ipam.IPRange{Start: "fd00::5", End: "fd00::9"},
-						&ipam.IPRange{Start: "fd00::18", End: "fd00::fffe"},
+					IPRangeList{
+						&IPRange{Start: "fd00::5", End: "fd00::9"},
+						&IPRange{Start: "fd00::18", End: "fd00::fffe"},
 					}))
 
 				Expect(subnet.V4NicToIP).To(BeEmpty())
@@ -647,43 +687,43 @@ var _ = Describe("[IPAM]", func() {
 			})
 
 			It("random allocation", func() {
-				subnet, err := ipam.NewSubnet(subnetName, "10.16.0.0/30,fd00::/126", nil)
+				subnet, err := NewSubnet(subnetName, "10.16.0.0/30,fd00::/126", nil)
 				Expect(err).ShouldNot(HaveOccurred())
 
 				ipv4, ipv6, _, err := subnet.GetRandomAddress("pod1.ns", "pod1.ns", "", nil, true)
 				Expect(err).ShouldNot(HaveOccurred())
-				Expect(ipv4).To(Equal(ipam.IP("10.16.0.1")))
-				Expect(ipv6).To(Equal(ipam.IP("fd00::1")))
+				Expect(ipv4).To(Equal(IP("10.16.0.1")))
+				Expect(ipv6).To(Equal(IP("fd00::1")))
 				ipv4, ipv6, _, err = subnet.GetRandomAddress("pod1.ns", "pod1.ns", "", nil, true)
 				Expect(err).ShouldNot(HaveOccurred())
-				Expect(ipv4).To(Equal(ipam.IP("10.16.0.1")))
-				Expect(ipv6).To(Equal(ipam.IP("fd00::1")))
+				Expect(ipv4).To(Equal(IP("10.16.0.1")))
+				Expect(ipv6).To(Equal(IP("fd00::1")))
 
 				ipv4, ipv6, _, err = subnet.GetRandomAddress("pod2.ns", "pod2.ns", "", nil, true)
 				Expect(err).ShouldNot(HaveOccurred())
-				Expect(ipv4).To(Equal(ipam.IP("10.16.0.2")))
-				Expect(ipv6).To(Equal(ipam.IP("fd00::2")))
+				Expect(ipv4).To(Equal(IP("10.16.0.2")))
+				Expect(ipv6).To(Equal(IP("fd00::2")))
 
 				_, _, _, err = subnet.GetRandomAddress("pod3.ns", "pod3.ns", "", nil, true)
-				Expect(err).Should(MatchError(ipam.ErrNoAvailable))
+				Expect(err).Should(MatchError(ErrNoAvailable))
 				Expect(subnet.V4FreeIPList).To(BeEmpty())
 				Expect(subnet.V6FreeIPList).To(BeEmpty())
 
-				Expect(subnet.V4IPToPod).To(HaveKeyWithValue(ipam.IP("10.16.0.1"), "pod1.ns"))
-				Expect(subnet.V4IPToPod).To(HaveKeyWithValue(ipam.IP("10.16.0.2"), "pod2.ns"))
-				Expect(subnet.V4NicToIP).To(HaveKeyWithValue("pod1.ns", ipam.IP("10.16.0.1")))
-				Expect(subnet.V4NicToIP).To(HaveKeyWithValue("pod2.ns", ipam.IP("10.16.0.2")))
-				Expect(subnet.V6IPToPod).To(HaveKeyWithValue(ipam.IP("fd00::1"), "pod1.ns"))
-				Expect(subnet.V6IPToPod).To(HaveKeyWithValue(ipam.IP("fd00::2"), "pod2.ns"))
-				Expect(subnet.V6NicToIP).To(HaveKeyWithValue("pod1.ns", ipam.IP("fd00::1")))
-				Expect(subnet.V6NicToIP).To(HaveKeyWithValue("pod2.ns", ipam.IP("fd00::2")))
+				Expect(subnet.V4IPToPod).To(HaveKeyWithValue(IP("10.16.0.1"), "pod1.ns"))
+				Expect(subnet.V4IPToPod).To(HaveKeyWithValue(IP("10.16.0.2"), "pod2.ns"))
+				Expect(subnet.V4NicToIP).To(HaveKeyWithValue("pod1.ns", IP("10.16.0.1")))
+				Expect(subnet.V4NicToIP).To(HaveKeyWithValue("pod2.ns", IP("10.16.0.2")))
+				Expect(subnet.V6IPToPod).To(HaveKeyWithValue(IP("fd00::1"), "pod1.ns"))
+				Expect(subnet.V6IPToPod).To(HaveKeyWithValue(IP("fd00::2"), "pod2.ns"))
+				Expect(subnet.V6NicToIP).To(HaveKeyWithValue("pod1.ns", IP("fd00::1")))
+				Expect(subnet.V6NicToIP).To(HaveKeyWithValue("pod2.ns", IP("fd00::2")))
 
 				subnet.ReleaseAddress("pod1.ns")
 				subnet.ReleaseAddress("pod2.ns")
-				Expect(subnet.V4FreeIPList).To(Equal(ipam.IPRangeList{}))
-				Expect(subnet.V4ReleasedIPList).To(Equal(ipam.IPRangeList{&ipam.IPRange{Start: "10.16.0.1", End: "10.16.0.2"}}))
-				Expect(subnet.V6FreeIPList).To(Equal(ipam.IPRangeList{}))
-				Expect(subnet.V6ReleasedIPList).To(Equal(ipam.IPRangeList{&ipam.IPRange{Start: "fd00::1", End: "fd00::2"}}))
+				Expect(subnet.V4FreeIPList).To(Equal(IPRangeList{}))
+				Expect(subnet.V4ReleasedIPList).To(Equal(IPRangeList{&IPRange{Start: "10.16.0.1", End: "10.16.0.2"}}))
+				Expect(subnet.V6FreeIPList).To(Equal(IPRangeList{}))
+				Expect(subnet.V6ReleasedIPList).To(Equal(IPRangeList{&IPRange{Start: "fd00::1", End: "fd00::2"}}))
 				Expect(subnet.V4IPToPod).To(BeEmpty())
 				Expect(subnet.V4NicToIP).To(BeEmpty())
 				Expect(subnet.V6IPToPod).To(BeEmpty())
