+++ pkg/ovs/ovn-nbctl-legacy_test.go
@@ -39,5 +39,5 @@ func Test_parseLrPolicyRouteListOutput(t *testing.T) {
         10                              ip4.src == 1.1.1.0/24            drop`
 	routeList, err := parseLrPolicyRouteListOutput(output)
 	ast.Nil(err)
-	ast.Equal(6, len(routeList))
+	ast.Equal(4, len(routeList))
 }
