--- test/unittest/unit_suite_test.go
@@ -1,17 +1,0 @@
-package unittest
-
-import (
-	"testing"
-
-	. "github.com/onsi/ginkgo/v2"
-	. "github.com/onsi/gomega"
-
-	// tests to run
-	_ "github.com/kubeovn/kube-ovn/test/unittest/ipam"
-	_ "github.com/kubeovn/kube-ovn/test/unittest/util"
-)
-
-func TestE2e(t *testing.T) {
-	RegisterFailHandler(Fail)
-	RunSpecs(t, "Kube-OVN unit test Suite")
-}
